Set PlatformTarget of ColorPicker, UDMFControls, TagExplorer and VisplaneExplorer to x86 (this should make Siberian_Tiger's work a little bit easier, I suppose...) VisplaneExplorer: fixed a crash when user tries to enable Visplane Explorer mode in a map created via New Map command.

+++ Source/Plugins/VisplaneExplorer/VisplaneExplorerMode.cs
@@ -389,7 +389,7 @@ namespace CodeImp.DoomBuilder.Plugins.VisplaneExplorer
 		}
 
 		// Processing
-		public override void OnProcess(double deltatime)
+		public override void OnProcess(float deltatime)
 		{
 			base.OnProcess(deltatime);
 			if(DateTime.Now >= nextupdate)
